requirements.txt
flask
google-generativeai
speechrecognition
pyttsx3
pyaudio
pygame
requests
